Add lazy-loaded resource properties.
This adds properties like SQS `queue.attributes` or S3 `obj.last_modified`,
which are lazy-loaded on access using the resource's `load` method. The
`load` and `reload` methods are added to facilitate this. Tests are added
to cover the new features.

One interesting case is e.g. SQS Message, which has **no** load method.
This object will (in the future) be loadable via SQS
`queue.receive_messages()` and allow access to e.g. `message.body`.
The work in this pull request is a prerequisit for that feature,
hence the test to ensure a load failure on such a resource by any other
means.

Also included is a really simple implementation of reverse JMESPath
support. Please ignore this for now, as it will be removed soon in favor
of the proper jmespath.create_structure call.