Refactor service action class and tests.
This commit refactors the `ServiceAction` class into several different
components and updates the associated tests. It pulls out the code
which builds request parameters, and the code which handles responses,
which greatly simplifies the actions. Testability is improved by
being significantly more readable and by limiting the required mocking
for a handful of tests.