Support latest resource model version.
This change adds support for the latest version of the JSON description files
for resources. The following has changed:

1. An action's `.path` has been moved into `.resource.path` if a resource
   is present. If it is present, then a path outside of the resource is
   no longer valid. This fixes #33.
2. Sub-resources previously defined with either `hasOne` or `hasMany` are now
   defined with a single `belongsTo` key. The internal Boto 3 representation,
   however, remains the same and is accessed via `resource.references`.

As a result of the format changes, I had to update the internal resource
model, adjust the search path behavior for actions and update various tests
that made assumptions about the older format.

I also discovered that we were using a mix of old and new files for different
services, so this change includes updates to EC2 and Glacier so that all
resources are using the new format. With this change, we no longer support
the older format.