Update resource model format.
This updates to the latest resource model format, which includes the following
changes:

1. `belongsTo` references are now named `has`
2. `subResources` is replaced by two-way `has` references
3. Identifier and parameter definitions are refactored:

    * `source` is now one of `name`, `path` or `value`
    * `sourceType` is now `source`
    * A new source of `input` is added for user-supplied input

Our interface remains the same, but the underlying JSON files have changed
and our factory code and documentation are somewhat simplified. The code
that handles identifiers/parameters is a bit more complex due to having
to check several keys in the dictionary based on the source.

The model still exposes subresources and references as separate concepts.
Due to circular references, resource instances now lazy-load the subresource
class when needed.

Tests and documentation have been updated to reflect these changes.