Load reference data if a resource path is defined
This change allows references with a JMESPath query set on the reference
resource path attribute to be loaded with data at instantiation time if
that data is present in the parent (via `meta.data`). If the data has
not yet been loaded and the parent is loadable, then a `load` operation
is incurred.

Before:
```python
>>> ni = ec2.NetworkInterface('abc123')
>>> ni.association.public_ip
ResourceLoadException: ec2.NetworkInterfaceAssociation has no load method
```

After:
```python
>>> ni = ec2.NetworkInterface('abc123')
>>> ni.association.public_ip
'127.0.0.1'
```

Added a test to ensure this works as expected.