Add nested structure descriptions to params & return value.
This change modifies the documentation to include structure examples for
both nested input parameters and nested service response values. Responses
are included for both clients and any resource action which returns a low-
level response.

This does **not** yet include documentation for each of the nested values
except for a basic Python type. I'm hoping to get some feedback on
this before including the full member documentation.

Additionally, this includes some basic tests for the new functionality
in the `docs.py` module. It also includes a minor naming change for
PEP8 (more of these are coming soon).