Remove page_size and limit from all
Removed the parameters to make the collections api less confusing by having
only one canonical way of limiting the number of results and sizes of pages.
Also ensures consistency of casing of arguments passed to method. Note that
for the filter() method, the limit and page_size arguments were just
undocumented to signify we do not officially support those parameters.