Fix double invocation of progress callbacks
Fixes #98.

This works by registering two handlers to disable/re-enable
callbacks during the request-created event handling.  So
for signature version 4, you now have::

    * request-created:
         - disable_callbacks() (via register_first())
         - client._sign_request()  (via register())
         - enable_callbacks()  (via register_last())

Added an integration test for this because the crux of this
working is because of the coordination of the event system
with the disabling/enabling of callbacks.

This will require a corresponding fix in botocore as well.