Add custom create_tag method to ec2 resource
The create tag method cannot be expressed via the resource model because
you need to map a set of tag to multiple resources possibly and return
a list of each tag for each resource. In order to preserve backward
combatability a new custom method needed to be added to the service
resource.