Add events property to boto3.session.Session
Before the only way to hook into the creating-client-class and
creating-resource-class events was to first create a botocore session
and pass that to the instantiation of the boto3.session.Session
object and register using the botocore session. This gives users access
to the event system through the boto3.session.Session object much
like you have access to it via the client.meta.events property.