Implement S3 copy transfer
Provide parallel multipart uploads for S3->S3 copies. The method is
named copy_file() to mirror upload_file() and download_file(). A new
MultipartCopier class is added that works similary to MultipartUploader
except that no osutil is required and a few extra arguments are
accepted for the upload_part_copy() calls.