Add Instance.delete_tags
The Tag resource was incorrectly modeled such that it is not a
collection on an Instance. As a result, there is no way to batch delete
tags from an Instance resource. This adds that functionality in, in a
way that won't break on future EC2 resource changes.

In the long term there are plans to codify SDK specific customizations
while maintaining a canonically correct resource model, but that will
sometime in the future as part of a different story.

Fixes boto/boto3#381