Ensure batch writer never sends more than flush_amount
In the case where no items are processed and we received
all the items back as unprocessed key, the next put_item
request will have flush_amount + 1, which will trigger
an error.

To fix this, the input_buffer is always sliced to
flush_amount before the batch_write_item() call.

Fixes #483.