Removed backcompat for OSUtils with callback
Realized that if open_file_chunk_reader() was ever invoked in the
TransferManager it would break because it would be providing a list of
callbacks. Furthermore, the method was never used so it did not make sense
to add a back compat especially since users should not be invoking the
method themselves. It should be up to the TransferManager to do that.