Added path attribute to the hasMany Stacks resource
The the cloudformations hasMany resources includes Stacks.
This is populated by the describe stacks call, but unlike other
resources of this type (i.e. ec2 instances) it does not populate the individual
stack resources with the results of this call.
The missing path attribute has been added to fix this.