Add LifecycleConfiguration resource
The BucketLifecycle resource uses the old GetBucketLifecycle
operation to load data. That operation only returns a single
transistion, even though there could be multiple transitions in a
single rule. This was fixed up by adding another command to the SDKs,
GetBucketLifecycleConfiguration. A new command was needed due to the
shift between a single object and a list of objects. The same change
is needed for the resource, and requires a new resource for the same
reason we needed a new command.

This adds the BucketLifecycleConfiguration resource in all the same
places that the BucketLifecycle resource existed. The resource itself
is identical in almost every way aside from which operations it
calls.