Unregister handlers when session is garbage collected
When boto3.session.Session registers a default handler, add a finalizer
to unregister the handler when the session is garbage collected. This
prevents excessive memory usage if the same botocore.session.Session
instance is reused by many instances of boto3.session.Session.

Because weakref.finalize is not available in Python 2, this fix is
implemented only for Python 3.

Fixes #2176