Fix DynamoDB BatchWriter dropping an item bug
I noticed that _items_buffer was set twice when BatchWriter flushed items:
first the buffer was moved past the items that were sent (according to _flush_amount),
and in the end it was emptied if sent items succeeded. This has a probably rare bug that
if a full batch of _flush_amount items was unprocessed, then the next item that triggers
flushing would be dropped as the buffer would have (_flush_amount + 1) items at that point
and the buffer would be emptied after sending the first _flush_amount items.

Added a test for the bug. Interestingly previous test "test_never_send_more_than_max_batch_size"
almost tested this but as it tested 2 unprocessed -> 1 unprocessed -> 0 unprocessed it didn't catch
this.

Also modified a little how the unprocessed items
are used so that logging number of unprocessed items is correct.